+++ copasi/UI/CQBaseDataModel.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2021 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -172,12 +172,12 @@ void CQBaseDataModel::fetchMore(const QModelIndex & parent)
   if (mFetchDisabled || parent.isValid())
     return;
 
-  int ToBeFetched = std::min(mFetchLimit, size() - mFetched);
+  int ToBeFetched = (int)std::min(mFetchLimit, size() - mFetched);
 
   if (ToBeFetched <= 0)
     return;
 
-  beginInsertRows(QModelIndex(), mFetched, mFetched + ToBeFetched - 1);
+  beginInsertRows(QModelIndex(), (int)mFetched, int(mFetched + ToBeFetched - 1));
 
   mFetched += ToBeFetched;
