+++ copasi/UI/CQCompartment.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2024 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -644,7 +644,7 @@ void CQCompartment::loadMetaboliteTable()
 {
   if (mpCompartment == NULL) return;
 
-  mpMetaboliteTable->setRowCount(mpCompartment->getMetabolites().size());
+  mpMetaboliteTable->setRowCount((int)mpCompartment->getMetabolites().size());
 
   CDataContainer::objectMap::const_iterator it =
     mpCompartment->getMetabolites().getObjects().begin();
